+++ wgengine/magicsock/magicsock_test.go
@@ -29,7 +29,7 @@ import (
 	"tailscale.com/types/key"
 	"tailscale.com/types/logger"
 	"tailscale.com/wgengine/filter"
-	"tailscale.com/wgengine/wgtun"
+	"tailscale.com/wgengine/tstun"
 )
 
 func TestListen(t *testing.T) {
@@ -328,9 +328,9 @@ func TestTwoDevicePing(t *testing.T) {
 	//t.Logf("cfg1: %v", uapi2)
 
 	tun1 := tuntest.NewChannelTUN()
-	wgtun1 := wgtun.WrapTUN(t.Logf, tun1.TUN())
-	wgtun1.SetFilter(filter.NewAllowAll())
-	dev1 := device.NewDevice(wgtun1, &device.DeviceOptions{
+	tstun1 := tstun.WrapTUN(t.Logf, tun1.TUN())
+	tstun1.SetFilter(filter.NewAllowAll())
+	dev1 := device.NewDevice(tstun1, &device.DeviceOptions{
 		Logger:         devLogger(t, "dev1"),
 		CreateEndpoint: conn1.CreateEndpoint,
 		CreateBind:     conn1.CreateBind,
@@ -343,9 +343,9 @@ func TestTwoDevicePing(t *testing.T) {
 	defer dev1.Close()
 
 	tun2 := tuntest.NewChannelTUN()
-	wgtun2 := wgtun.WrapTUN(t.Logf, tun2.TUN())
-	wgtun2.SetFilter(filter.NewAllowAll())
-	dev2 := device.NewDevice(wgtun2, &device.DeviceOptions{
+	tstun2 := tstun.WrapTUN(t.Logf, tun2.TUN())
+	tstun2.SetFilter(filter.NewAllowAll())
+	dev2 := device.NewDevice(tstun2, &device.DeviceOptions{
 		Logger:         devLogger(t, "dev2"),
 		CreateEndpoint: conn2.CreateEndpoint,
 		CreateBind:     conn2.CreateBind,
@@ -391,7 +391,7 @@ func TestTwoDevicePing(t *testing.T) {
 	t.Run("ping 1.0.0.2", func(t *testing.T) { ping2(t) })
 	t.Run("ping 1.0.0.2 via SendPacket", func(t *testing.T) {
 		msg1to2 := tuntest.Ping(net.ParseIP("1.0.0.2"), net.ParseIP("1.0.0.1"))
-		if err := wgtun1.InjectOutbound(msg1to2); err != nil {
+		if err := tstun1.InjectOutbound(msg1to2); err != nil {
 			t.Fatal(err)
 		}
 		select {
