+++ wgengine/watchdog_test.go
@@ -12,7 +12,7 @@ import (
 	"time"
 
 	"tailscale.com/wgengine/router"
-	"tailscale.com/wgengine/wgtun"
+	"tailscale.com/wgengine/tstun"
 )
 
 func TestWatchdog(t *testing.T) {
@@ -20,7 +20,7 @@ func TestWatchdog(t *testing.T) {
 
 	t.Run("default watchdog does not fire", func(t *testing.T) {
 		t.Parallel()
-		tun := wgtun.WrapTUN(t.Logf, wgtun.NewFakeTUN())
+		tun := tstun.WrapTUN(t.Logf, tstun.NewFakeTUN())
 		e, err := NewUserspaceEngineAdvanced(t.Logf, tun, router.NewFake, 0)
 		if err != nil {
 			t.Fatal(err)
@@ -37,7 +37,7 @@ func TestWatchdog(t *testing.T) {
 
 	t.Run("watchdog fires on blocked getStatus", func(t *testing.T) {
 		t.Parallel()
-		tun := wgtun.WrapTUN(t.Logf, wgtun.NewFakeTUN())
+		tun := tstun.WrapTUN(t.Logf, tstun.NewFakeTUN())
 		e, err := NewUserspaceEngineAdvanced(t.Logf, tun, router.NewFake, 0)
 		if err != nil {
 			t.Fatal(err)
